an example of how diststore could be used for assets
The primary change here is to add a concept of extra
stores to the plugin's config and then in `init` modify
whatever the current store is to be a diststore using
those extras.

The names are probably not correct. They are presented here
to show the concept rather than be complete. In fact, the whole
thing is shown to demonstate the basics of the concept. The
details might turn out to be different.

A few caveats:

* If this style is used, then bfwinstance will no longer be correct as
  `spawn` will install a store_structure and copy over store_contents.
  If this is gonna be a thing a different instance script may be desired
  OR ibuilder should perhaps take a flag that says "only make the dir
  and write the config".

* If the store being used is the caching store, then whenever the
  package store contents are updated memcached will need to be restarted
  as there will never be any successful writes to the package store (it
  is read only) to trigger the hooks which do cache invalidation.

Note the XXX about moving some config items to config.py